+++ src/krux/wallet.py
@@ -24,7 +24,7 @@ from embit.descriptor.arguments import Key
 from embit.networks import NETWORKS
 from embit.bip32 import HARDENED_INDEX
 from .krux_settings import t
-from .qr import FORMAT_BBQR
+from .qr import FORMAT_BBQR, FORMAT_NONE
 from .key import (
     P2PKH,
     P2SH_P2WPKH,
@@ -49,7 +49,7 @@ class Wallet:
     def __init__(self, key):
         self.key = key
         self.wallet_data = None
-        self.wallet_qr_format = None
+        self.wallet_qr_format = FORMAT_NONE
         self.descriptor = None
         self.label = None
         self.policy = None